re-organized files, removed debug statement
